Included additional utilities in Manager Base
