+++ src/storage/managers/base.pyi
@@ -26,6 +26,10 @@ class BaseManagerInterface(ABC, Generic[ManagerT]):
 
     def _is_static(self) -> bool: ...
 
+    def _get_identifier_column_name(self) -> str: ...
+
+    def _get_identifier(self) -> Identifier: ...
+
     def save(self) -> None: ...
 
     def update(self, data: Dict[str, Any]) -> DatabaseObject: ...
